requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-groq
tavily-python